Refactored main module to several submodules under tradeexecutor.cli.commands
