+++ tradeexecutor/cli/commands/hello.py
@@ -1,0 +1,7 @@
+from tradeexecutor.cli.commands.app import app
+
+
+@app.command()
+def hello():
+    """Check that the application loads without doing anything."""
+    print("Hello blockchain")
